+++ src/Dialogs/dlgWayPointDetails.cpp
@@ -468,7 +468,7 @@ dlgWayPointDetailsShowModal(SingleWindow &parent, const Waypoint& way_point)
   wp = (WndProperty *)wf->FindByName(_T("prpMc2"));
   if (wp) {
     _stprintf(sTmp, _T("%.0f %s"),
-              Units::ToUserAltitude(r.AltitudeDifference),
+              (double)Units::ToUserAltitude(r.AltitudeDifference),
               Units::GetAltitudeName());
     wp->SetText(sTmp);
   }
@@ -480,7 +480,7 @@ dlgWayPointDetailsShowModal(SingleWindow &parent, const Waypoint& way_point)
   wp = (WndProperty *)wf->FindByName(_T("prpMc0"));
   if (wp) {
     _stprintf(sTmp, _T("%.0f %s"),
-              Units::ToUserAltitude(r.AltitudeDifference),
+              (double)Units::ToUserAltitude(r.AltitudeDifference),
               Units::GetAltitudeName());
     wp->SetText(sTmp);
   }
@@ -491,7 +491,7 @@ dlgWayPointDetailsShowModal(SingleWindow &parent, const Waypoint& way_point)
   wp = (WndProperty *)wf->FindByName(_T("prpMc1"));
   if (wp) {
     _stprintf(sTmp, _T("%.0f %s"),
-              Units::ToUserAltitude(r.AltitudeDifference),
+              (double)Units::ToUserAltitude(r.AltitudeDifference),
               Units::GetAltitudeName());
     wp->SetText(sTmp);
   }
